cleans up components with default layout
