+++ pages/icons.js
@@ -4,24 +4,18 @@
 import Head from "next/head";
 import Link from 'next/link';
 import React, { useEffect, useState } from "react";
-import * as icons from '@wordpress/icons';
 
 /**
  * Internal Dependencies
  */
 import Footer from "@components/Footer";
-import IconPreview from "@components/IconPreview";
-
+import IconGrid from "@components/IconGrid";
 
 
 
 export default function Icons() {
 
-	const [ search, setSearch ] = useState('');
-
-	const handleSearch = (event) => {
-		setSearch(event.target.value);
-	}
+	
 
 	return(
 		<div className="container">
@@ -72,20 +66,7 @@ export default function Icons() {
 			</Head>
 
 			<main>
-				<header className="icons-header">
-					<h1>WordPress Icons Library.</h1>
-					<p>A searchable guide to the complete <a href="https://github.com/WordPress/gutenberg/tree/trunk/packages/icons" target="_blank">WordPress Icon Library</a> package from Gutenberg.</p>
-					<input type="text" placeholder="Search" value={search} onChange={handleSearch} />
-				</header>
-				<section className="icons-grid">
-					{
-						Object.keys(icons).filter((icon)=>{ return search === '' || icon.toLowerCase().indexOf(search.toLowerCase()) > -1 }).map((icon, index) => {
-							return  ( 'Icon' !== icon ) ? (
-								<IconPreview icon={ icon } name={ icons[icon] } key={index} />
-							) : null;
-						}
-					)}
-				</section>
+				<IconGrid />
 			</main>
 
 			<Footer />
